+++ autocnet/io/db/connection.py
@@ -1,14 +1,8 @@
-import socket
-
-import sqlalchemy
-from sqlalchemy import create_engine, pool, orm
-from sqlalchemy.orm import create_session, scoped_session, sessionmaker
-
 import logging
-import os
 import socket
-import warnings
-import yaml
+from time import sleep
+
+from sqlalchemy import orm, create_engine, pool
 
 # set up the logging file
 log = logging.getLogger(__name__)
@@ -19,7 +13,24 @@ class Parent:
         self.session = Session()
         self.session.begin()
 
-def new_connection(dbconfig):
+def retry(max_retries=5, wait_time=300):
+    def decorator(func):
+        def wrapper(*args, **kwargs):
+            retries = 0
+            if retries < max_retries:
+                try:
+                    result = func(*args, **kwargs)
+                    return result
+                except:
+                    retries += 1
+                    sleep(wait_time)
+            else:
+                raise Exception(f"Maximum retries of {func} exceeded! Is the database accessible?")
+        return wrapper
+    return decorator
+
+@retry()
+def new_connection(dbconfig, with_session=False):
     """
     Using the user supplied config create a NullPool database connection.
 
@@ -29,6 +40,8 @@ def new_connection(dbconfig):
                Dictionary defining necessary parameters for the database
                connection
 
+    with_session : boolean
+                   If true return a SQL Alchemy session factory. Default False.
     Returns
     -------
     Session : object
@@ -43,11 +56,12 @@ def new_connection(dbconfig):
                                                   dbconfig['pgbouncer_port'],
                                                   dbconfig['name'])
     hostname = socket.gethostname()
-    engine = sqlalchemy.create_engine(db_uri,
-                poolclass=sqlalchemy.pool.NullPool,
+    engine = create_engine(db_uri,
+                poolclass=pool.NullPool,
                 connect_args={"application_name":f"AutoCNet_{hostname}"},
-                isolation_level="AUTOCOMMIT",
                 pool_pre_ping=True)
-    Session = orm.sessionmaker(bind=engine, autocommit=False)
-    log.debug(Session, engine)
-    return Session, engine
+    if with_session:
+        Session = orm.sessionmaker(bind=engine, autocommit=False)
+        log.debug(Session, engine)
+        return Session, engine
+    return engine