+++ autocnet/matcher/subpixel.py
@@ -9,22 +9,20 @@ import warnings
 import logging
 
 import numpy as np
-import pvl
 from scipy.ndimage import center_of_mass
 from skimage import transform as tf
 from skimage import registration
 from skimage import filters
 from scipy import fftpack
 from scipy.spatial import distance_matrix
-from sqlalchemy.sql.expression import bindparam
 from sqlalchemy import inspect
 from matplotlib import pyplot as plt
 
-from autocnet.matcher.naive_template import pattern_match, pattern_match_autoreg
+from autocnet.matcher.naive_template import pattern_match
 from autocnet.matcher.mutual_information import mutual_information
-from autocnet.io import isis
 from autocnet.io.geodataset import AGeoDataset
 from autocnet.io.db.model import Measures, Points, Images, JsonEncoder
+from autocnet.io.db.utils import get_point, get_nodes_for_measures, update_measures, ignore_measures
 from autocnet.graph.node import NetworkNode
 from autocnet.transformation import roi
 from autocnet.transformation.affine import estimate_local_affine
@@ -807,6 +805,7 @@ def subpixel_register_point_smart(point,
                                   chooser='subpixel_register_point_smart',
                                   verbose=False,
                                   func=pattern_match,
+                                  maximum_affine_reprojective_error=1.0,
                                   ncg=None):
 
     """
@@ -837,40 +836,27 @@ def subpixel_register_point_smart(point,
                  {'match_kwargs': {'image_size':(151,151), 'template_size':(67,67)}},
                  {'match_kwargs': {'image_size':(181,181), 'template_size':(73,73)}}]
     """
-    with ncg.session_scope() if ncg is not None else nullcontext(session) as session:
-        if not isinstance(point, Points):
-            point = session.query(Points).filter(Points.id == point).one()
-        pointid = point.id
+    if not isinstance(point, Points):
+        point = get_point(ncg, session, point)
+    pointid = point.id
+    measures = point.measures
+    reference_index = point.reference_index
+    
+    # measures = session.query(Measures).filter(Measures.pointid == pointid).order_by(Measures.id).all()
 
-        # Order by is important here because the measures get ids in sequential order when initially placed
-        # and the reference_index is positionally linked to the ordered vector of measures.
-        measures = session.query(Measures).filter(Measures.pointid == pointid).order_by(Measures.id).all()
-        reference_index = point.reference_index
+    # Get the reference measure to instantiate the source node. All other measures will
+    # match to the source node.
+    source = measures[reference_index]
+    reference_index_id = source.imageid
+    log.info(f'Using measure {source.id} on image {source.imageid}/{source.serial} as the reference.')
+    log.info(f'Measure reference index is: {reference_index}')
 
-        # Get the reference measure to instantiate the source node. All other measures will
-        # match to the source node.
-        source = measures[reference_index]
-        reference_index_id = source.imageid
-        log.info(f'Using measure {source.id} on image {source.imageid}/{source.serial} as the reference.')
-        log.info(f'Measure reference index is: {reference_index}')
-
-        # Build a node cache so that this is an encapsulated database call. Then nodes
-        # can be pulled from the lookup sans database.
-        nodes = {}
-        for measure in measures:
-            res = session.query(Images).filter(Images.id == measure.imageid).one()
-            logging.debug(f'Node instantiation image query result: {res.path, res.cam_type, res.dem, res.dem_type}')
-            nn = NetworkNode(node_id=measure.imageid, 
-                            image_path=res.path,
-                            cam_type=res.cam_type,
-                            dem=res.dem,
-                            dem_type=res.dem_type)
-            nodes[measure.imageid] = nn
-            session.expunge_all()
-
+    # Build a node cache so that this is an encapsulated database call. Then nodes
+    # can be pulled from the lookup sans database.
+    nodes_cache = get_nodes_for_measures(ncg, session, measures)
     log.info(f'Attempting to subpixel register {len(measures)-1} measures for point {pointid}')
     # Set the reference image
-    source_node = nodes[reference_index_id]
+    source_node = nodes_cache[reference_index_id]
 
     log.info(f'Source: sample: {source.sample} | line: {source.line}')
     updated_measures = []
@@ -880,20 +866,25 @@ def subpixel_register_point_smart(point,
             continue
 
         cost = None
-        destination_node = nodes[measure.imageid]
+        destination_node = nodes_cache[measure.imageid]
         log.info(f'Registering measure {measure.id} (image: {measure.imageid}, serial: {measure.serial})')
 
+        match_kwargs = parameters[0]['match_kwargs']
         reference_roi = roi.Roi(source_node.geodata, 
                                 source.apriorisample, 
-                                source.aprioriline)
+                                source.aprioriline,
+                                size_x=match_kwargs['image_size'][0],
+                                size_y=match_kwargs['image_size'][1])
 
         moving_roi = roi.Roi(destination_node.geodata, 
                              measure.apriorisample, 
                              measure.aprioriline,
-                             )
+                             size_x=match_kwargs['image_size'][0],
+                             size_y=match_kwargs['image_size'][1])
         try:
             baseline_affine = estimate_local_affine(moving_roi,
-                                                    reference_roi)
+                                                    reference_roi,
+                                                    maximum_affine_reprojective_error)
         except Exception as e:
             log.exception(e)
             m = {'id': measure.id,
@@ -907,8 +898,13 @@ def subpixel_register_point_smart(point,
         # If the image read center plus buffer is outside the image, clipping
         # raises an index error. Handle and set the measure false.
         try:
-            reference_clip = reference_roi.clip(affine=baseline_affine, buffer=10)
-            moving_clip = moving_roi.clip(buffer=5)
+            reference_clip = reference_roi.clip(affine=baseline_affine, 
+                                                buffer=10,
+                                                size_x=match_kwargs['image_size'][0],
+                                                size_y=match_kwargs['image_size'][1])
+            moving_clip = moving_roi.clip(buffer=5,
+                                          size_x=match_kwargs['image_size'][0],
+                                          size_y=match_kwargs['image_size'][1])
         except Exception as e:
             log.error(e)
             m = {'id': measure.id,
@@ -993,12 +989,13 @@ def subpixel_register_point_smart(point,
                 'template_metric':maxcorr,
                 'template_shift':dist,
                 'mi_metric': 0,
-                'status': True}
+                'status': True,
+                'imageid':measure.imageid}
             log.info(f'METRIC: {maxcorr}| SAMPLE: {new_x} | LINE: {new_y} | MI: 0')
             
             updated_measures.append([baseline_mi, baseline_corr, m])
     # Baseline MI, Baseline Correlation, updated measures to select from
-    return updated_measures
+    return updated_measures, nodes_cache
 
 def check_for_shift_consensus(shifts, tol=0.1):
     """
@@ -1064,6 +1061,9 @@ def decider(measures, tol=0.6):
                             of meaure ids to be ignored beause theu fail the consensus
                             building approach
     """
+    #TODO: This is super janky. I had to add imageid above, int he func, to the by_id
+    # list, and then again to the else statement at the end. This is way to hard
+    # to extend with new attributes.
     by_id = defaultdict(list)
     measures_to_set_false = []
     for m in measures:
@@ -1078,7 +1078,8 @@ def decider(measures, tol=0.6):
                                   m['template_metric'],
                                   baseline_mi,
                                   baseline_corr,
-                                  m['template_shift']])
+                                  m['template_shift'],
+                                  m['imageid']])
         else:
             measures_to_set_false.append(m['id'])
 
@@ -1107,7 +1108,8 @@ def decider(measures, tol=0.6):
                  'template_shift': best_measure[6],
                  'choosername': choosername,
                  'ignore':False,
-                 'best_parameter_index': best_cost}
+                 'best_parameter_index': best_cost,
+                 'imageid': best_measure[7]}
             measures_to_update.append(m)
     # A measure could have one bad regitration and get set false, if a different parameter set passed,
     # remove from the set false list.
@@ -1116,11 +1118,12 @@ def decider(measures, tol=0.6):
 
     return measures_to_update, measures_to_set_false
 
-def validate_candidate_measure(measure_to_register,
-                               session=None,
-                               ncg=None,
+def validate_candidate_measure(point,
+                               measure_to_register,
+                               node_cache,
                                parameters=[],
                                func=pattern_match,
+                               maximum_affine_reprojective_error=1.0,
                                **kwargs):
     """
     Compute the matching distances, matching the reference measure to the measure
@@ -1158,45 +1161,20 @@ def validate_candidate_measure(measure_to_register,
             Of reprojection distances for each parameter set.
     """
 
-    measure_to_register_id = measure_to_register['id']
-
-    with ncg.session_scope() if ncg is not None else nullcontext(session) as session:
-        # Get the measure to be registered
-        measure = session.query(Measures).filter(Measures.id == measure_to_register_id).order_by(Measures.id).one()
-
-        # Get the references measure
-        point = measure.point
-        reference_index = point.reference_index
-        reference_measure = point.measures[reference_index]
-
-
-        # Match the reference measure to the measure_to_register - this is the inverse of the first match attempt
-        # Source is the image that we are seeking to validate, destination is the reference measure.
-        # This is the inverse of other functions as this is a validator.
-
-        source_imageid = measure.imageid
-        source_image = session.query(Images).filter(Images.id == source_imageid).one()
-        source_node = NetworkNode(node_id=source_imageid, 
-                                  image_path=source_image.path, 
-                                  cam_type=source_image.cam_type,
-                                  dem=source_image.dem,
-                                  dem_type=source_image.dem_type)
-        source_node.parent = ncg
-
-        destination_imageid = reference_measure.imageid
-        destination_image = session.query(Images).filter(Images.id == destination_imageid).one()
-        destination_node = NetworkNode(node_id=destination_imageid, 
-                                       image_path=destination_image.path,
-                                       cam_type=source_image.cam_type,
-                                       dem=destination_image.dem,
-                                       dem_type=destination_image.dem_type)
-        destination_node.parent = ncg
-        session.expunge_all()
+    
+    # Get the source (being validation) and destination (current reference measure)
+    # from the node_cache
+    source_node = node_cache[measure_to_register['imageid']]
+    reference_index = point.reference_index
+    reference_measure = point.measures[reference_index]
+    destination_node = node_cache[reference_measure.imageid]
 
     sample = measure_to_register['sample']
     line = measure_to_register['line']
 
-    log.info(f'Validating measure: {measure_to_register_id} on image: {source_imageid}')
+    mid = measure_to_register['id']
+    sid = measure_to_register['imageid']
+    log.info(f'Validating measure: {mid} on image: {sid}')
 
     reference_roi = roi.Roi(source_node.geodata, 
                             sample, 
@@ -1208,7 +1186,9 @@ def validate_candidate_measure(measure_to_register,
                             buffer=20)
     
     try:
-        baseline_affine = estimate_local_affine(moving_roi, reference_roi)
+        baseline_affine = estimate_local_affine(moving_roi, 
+                                                reference_roi,
+                                                maximum_affine_reprojective_error)
     except:
        log.error('Unable to transform image to reference space. Likely too close to the edge of the non-reference image. Setting ignore=True')
        return [np.inf] * len(parameters)
@@ -1245,6 +1225,7 @@ def smart_register_point(point,
                          parameters=[], 
                          shared_kwargs={}, 
                          valid_reprojection_distance=1.5, 
+                         maximum_affine_reprojective_error=1.0,
                          ncg=None):
     """
     The entry func for the smart subpixel registration code. This is the user
@@ -1281,6 +1262,12 @@ def smart_register_point(point,
                                   measures matched from the moving image to the reference image with a 
                                   distance less than this value in pixels are considered valid. Default: 1.1 
 
+    maximum_affine_reprojective_error: float
+                                       reprojective errors less than or equal to this value (default 1.0) are
+                                       masked and removed from affine transform estimations. If less than 3 points
+                                       are found, affine estimation will fail. This value can be used to loosen
+                                       the reprojective constraint estimating the affine transform between 
+                                       images.
     Returns
     -------
     measures_to_update : list
@@ -1291,17 +1278,25 @@ def smart_register_point(point,
                             building approach
 
     """
-    with ncg.session_scope() if ncg is not None else nullcontext(session) as session:
-        if not isinstance(point, Points):
-            point = session.query(Points).filter(Points.id == point).one()
-        session.expunge_all()
+    if not isinstance(point, Points):
+        point = get_point(ncg, session, point)
         
-    measure_results = subpixel_register_point_smart(point, session, parameters=parameters, **shared_kwargs)
+    measure_results, node_cache = subpixel_register_point_smart(point, 
+                                                                session, 
+                                                                ncg=ncg, 
+                                                                parameters=parameters, 
+                                                                maximum_affine_reprojective_error=maximum_affine_reprojective_error
+                                                                **shared_kwargs)
     measures_to_update, measures_to_set_false = decider(measure_results)
-    log.info(f'Found {len(measures_to_update)} measures that found subpixel registration consensus. Running validation now...')
+    log.info(f'Found {len(measures_to_update)} measures that found subpixel registration consensus.')
     # Validate that the new position has consensus
     for measure in measures_to_update:
-        reprojection_distances = validate_candidate_measure(measure, session, parameters=parameters, ncg=ncg, **shared_kwargs)
+        reprojection_distances = validate_candidate_measure(point, 
+                                                            measure, 
+                                                            node_cache, 
+                                                            parameters=parameters, 
+                                                            maximum_affine_reprojective_error=maximum_affine_reprojective_error,
+                                                            **shared_kwargs)
         log.info(f'Validation Distance Boolean: {np.array(reprojection_distances) < valid_reprojection_distance}')
         if np.sum(np.array(reprojection_distances) < valid_reprojection_distance) < 2:
             log.info(f"Measure {measure['id']} failed validation. Setting ignore=True for this measure.")
@@ -1310,28 +1305,9 @@ def smart_register_point(point,
     for measure in measures_to_update:
         measure['_id'] = measure.pop('id', None)
 
-
-    # Update the measures that passed registration
-    if measures_to_update:
-        stmt = Measures.__table__.update().\
-                                where(Measures.__table__.c.id == bindparam('_id')).\
-                                values({'weight':bindparam('weight'),
-                                        'measureIgnore':bindparam('ignore'),
-                                        'templateMetric':bindparam('template_metric'),
-                                        'templateShift':bindparam('template_shift'),
-                                        'line': bindparam('line'),
-                                        'sample':bindparam('sample'),
-                                        'ChooserName':bindparam('choosername')})
-        session.execute(stmt, measures_to_update)
-
-    if measures_to_set_false:
-        measures_to_set_false = [{'_id':i} for i in measures_to_set_false]
-        # Set ignore=True measures that failed
-        stmt = Measures.__table__.update().\
-                                where(Measures.__table__.c.id == bindparam('_id')).\
-                                values({'measureIgnore':True,
-                                        'ChooserName':shared_kwargs['chooser']})
-        session.execute(stmt, measures_to_set_false)
+    # Update the measures that passed and failed registration
+    update_measures(ncg, session, measures_to_update)
+    ignore_measures(ncg, session, measures_to_set_false, shared_kwargs['chooser'])
 
     log.info(f'Updated measures: {json.dumps(measures_to_update, indent=2, cls=JsonEncoder)}')
     log.info(f'Ignoring measures: {measures_to_set_false}')